+++ ticketguardian/client/client.py
@@ -2,10 +2,16 @@ from ticketguardian.affiliate import Affiliate
 from ticketguardian.abstract import (
     ListResourceMixin,
     PostResourceMixin,
+    PatchResourceMixin,
     RetrieveResourceMixin, )
 
 
-class Client(ListResourceMixin, RetrieveResourceMixin, PostResourceMixin, ):
+class Client(
+        ListResourceMixin,
+        PatchResourceMixin,
+        PostResourceMixin,
+        RetrieveResourceMixin,):
+
     resource = "clients"
 
     @property
@@ -17,3 +23,12 @@ class Client(ListResourceMixin, RetrieveResourceMixin, PostResourceMixin, ):
     @property
     def domain(self):
         return
+
+    @property
+    def scope(self):
+        """
+        Used to gets all children within the scope of this Client.
+        Returns:
+            list: a list of all children of the Client including itself.
+        """
+        return self.retrieve(self.id, 'scope', raw_data=True)