+++ ticketguardian/order/tests/test_charge_order.py
@@ -62,7 +62,18 @@ def test_charge_order():
     }
 
     for field in charge_fields:
+        is_unicode = False
         assert hasattr(charge, field)
-        assert isinstance(getattr(charge, field), charge_fields[field])
+        if charge_fields[field] == str:
+            try:
+                # A little hacky but unicode is not defined in Python3
+                # so if this fails then we are testing Python2
+                is_unicode = isinstance(getattr(charge, field), unicode)
+            except NameError:
+                pass
+
+        assert is_unicode or isinstance(
+            getattr(charge, field), charge_fields[field]
+        )
 
     return charge